move jsonutils to lib
